Proposal: use %rsp for the stack pointer on x86_64
This came up a couple of times recently, and I did a bit of digging to see whether it might be a good idea.  Summary: I suspect it's not going to buy us very much and might even be a loss performance-wise due to the increase in code size, but at least this proposal will serve as a place to collect the information.