Give reify1## a type that will work in Core
We need to be able to instantiate the constrained function at
a type the specializer knows about, but get out a type the user
asks for. My previous guess at the method type would have needed
an `unsafeCoerce#` to finish things off; no good! So the new
version takes an "unwrapper" argument to do that job.