QualifiedImports
This proposal adds a new extension to GHC `-XQualifiedImports` which
switches the default behavior of `import` from importing modules
unqualified to importing module qualified. It also introduces a notion of *main type* of a module import: the main type of a module import is automatically imported unqualified.

This proposal is authored by @aspiwack and @guibou.

[Rendered](https://github.com/tweag/ghc-proposals/blob/qualified-import/proposals/0000-default-qualified-import.rst)