th-stage-hygiene: New "effects" subsection on build speed
Cover some interactions there, stemming from lighter depedencies and the
equational reasoning of stages. Rearrange treatment of core way and
naive core interpreter as they are purely optional and so move to that
section. Also clarify that plain hi is fine for stage 1, but "fat"
interface file (thanks @michaelpj for the tip!) are needed for stage -1.

The GHC subsection now comes after that, so it can say less on the naive
core interpreter, and more on the big picture.

Capitalize "Core".