Remove sectnum
this was put ther a while ago to make the sphinx output have
non-consequtive section numbers that match the proposal number, e.g. on
https://ghc-proposals.readthedocs.io/en/latest/

This is nice, but
 * it is not _crucial_ for that output
 * it kinda breaks the rendering of proposals on GitHub

People keep asking me about this, to time to fix this.

If we ever make the sphinx output “official” we can revise that.

Fixes issue #298.