Remove “proposal number” from metadata
it was wrong for most proposals since
33bf34151510313ae70572a1a103093e4cd65b93 and nobody seemed to notice or
care.

Also less noise, less to do when merging and a shorter header is good.

The proposal number is easily visible in the
filename, and also from the link to the original PR.

I can still add them back when and if someone notices and cares.