Adjust syntax for linear arrows
This is not a full proposal. It's a small change to a recently
accepted proposal, ahead of the implementation making it into
a release.

In the linear types proposal as accepted, we proposed to introduce
a new type constructor `FUN p a b` where `p :: Multiplicity`, along
with *two* new syntactic forms in types:

* `a #p -> b` stands for `FUN p a b`;
* because multiplicity `One` is so common, `a #-> b` stands for `FUN
  'One a b`.

Based on implementation experience writing `linear-base`, this change
removes the second syntactic form. The motivations are:

* From a didactic point of view, we prefer to convey the idea that
  there really is only one function space in Haskell, denoted `a ->
  b`. The arrow is parameterized in various ways, including by
  a multiplicity. The `a #p -> b` syntax is in keeping with this idea.
* `#->` is 3 characters long where `::` and `->` are of width 2. This
  makes aligning a mix of all three lexemes vertically in a type
  signature awkward.

We propose to henceforth denote a linear function with type `a #1 -> b`.
For now, this is technically a new syntactic form that replaces `a #-> b`.
In the future, this may or may not remain special. If the type-level
is further unified with the term-level as some have plans for, it is
possible that integer literals in types will become overloaded, like
they are in terms. If that happens, we'll have `1:: GHC.Types.Nat` or
`1 :: Multiplicity` depending on context, and `a #1 -> b` will no
longer be a special syntactic form. Anything between between `#` and
`->` will then truly be an arbitrary type in all cases.

We have tentative plans for other multiplicities in the future, like
multiplicity `0`.

Overall, we view this change to the syntax of linear arrows as
a simplification: the user need only learn one new syntactic form, not
two.