Add committee bylaws (#360)
A meeting of the committee at ICFP led to an understanding that we need to codify our practices a bit. This is, thankfully, not a result of anything having gone wrong, but an agreement that our continued success requires a well-articulated foundation.

This PR is a stab at drafting bylaws. The committee did *not* discuss particulars, which are my own invention, and about which I expect suitable debate.

Note that this PR is set up to allow edits by the committee members; please feel free to edit my branch directly, especially with typos and formatting goofs.

[Rendered bylaws](https://github.com/ghc-proposals/ghc-proposals/blob/bylaws/committee.rst), but there are a few other changes with this PR as well.