no-implicit-binds: Adjust defaulting of `-XPatternSignatureBinds`
Today `-XPatternSignatureBinds` only matters when `-XScopeTypeVariables`
is enabled, but proposal #238 would re-introduce `-XPatternSignatures`,
with which `-XPatternSignatureBinds` would also be visible.

We cannot change the meaning of `-XScopeTypeVariables`, but I would
rather `-XPatternSignatures` alone not cause any variables to be bound,
as this is the more conservative, forwards compatible option, and one
that I believe would best position `-XPatternSignatures` to become part
of a future Haskell report even if the other extensions are hotly
debated.

Finally, this means #238 can be a bit simpler and just assume the
existence of `-XPatternSignatureBinds` without having to change any
aspect of it.