Support multiple hole meanings in types
Holes in a type can either mean that the user does not know what belongs
there or that they don't care. At present, this can only be controlled
for type signatures at the module level, and not at all for type
applications. I propose a simple way to distinguish these holes
syntactically.

[Rendered proposal](https://github.com/treeowl/ghc-proposals/blob/different-holes/proposals/0000-distinguish-partial-sigs.md)