Type arguments in constructor patterns: fix signature scoping
At the term-level, we reject nonlinear variable bindings:
	k (x, x) = ()
	-- Error: Conflicting definitions for ‘x’

However, pattern signatures are exempt from this restriction:
	u (Nothing :: Maybe (x, x)) = ()

Now consider the following program:
	data Proxy a = P
	g :: Proxy (Nothing @(a, a)) -> ()
	g (P @(Nothing :: Maybe (x, x))) = ()

Should we accept or reject ‘g’? Is it more like ‘u’ or like ‘k’?
Since multiple occurrences of ‘x’ are on the right-hand side of ::,
one might expect it to be treated like ‘u’. Unfortunately, the
proposal fails to take this into account and states the following:

	Any type variables mentioned in a type application are considered
	binding sites, shadowing any in-scope type variables.

Presumably, "any type variables" includes type variables in pattern
signatures, so we would end up rejecting ‘g’. This, however, is in
direct contradiction with the proposal's stated goal to uphold the VOP
(Visibility Orthogonality Principle).

The fact that the pattern signature in ‘g’ is hidden under a ``@``
should not affect its scoping rules! This amendment corrects this
oversight and restores the VOP.