Adjust warning category syntax
This adjusts the syntax of warning categories in #541 so that it is easier to
parse and less ambiguous (both for computers and humans). The original proposal
allowed a module to be annotated with

    {-# WARNING [x-category] "message" #-}
    {-# WARNING [x-category] ["message1", "message2"] #-}

with square brackets used for both category and a list of message strings.
This changes the syntax to use the `in` keyword and a string instead, like this:

    {-# WARNING in "x-category" "message" #-}
    {-# WARNING in "x-category" ["message1", "message2"] #-}