HasField redesign
This proposal recommends improvements to the design of the
`HasField` typeclass (differing from those planned under #158).
In particular, the proposed design supports updates using a new
`SetField` class, adds support for unlifted datatypes and fields,
and specifies laws for the classes.

In order to keep this proposal simple, it does not yet propose
adding support for type-changing update, which is left for a
future proposal.