Amend #522
This is an amendment to proposal #522.

We propose three changes to the current `one of` syntax:
 1. change syntax from `one of 1, 2, ...` to `one of {1; 2; ...}`. We don't require the curly braces however, so `one of 1;2;...` is also good.
 2. put production into `pat` instead of `apat` and without parentheses around it
 3. allow or-pattern to have 0 or 1 arguments.

Why?

### 1. `one of {...}` syntax

Similar to `case of`, we want to make use of layout blocks. Writing explicit `{}` around the pattern gives a much nicer grouping for code readers. Previously, one could misread `(one of 1, 2)` as `((one of 1), 2)`. Now, `one of {1;2}` makes this impossible.

Still, the `{}` are optional. This allows us to write or-patterns in a multilined fashion:
```hs
f (one of 1
               2
               3) -> ...
```

`of` layout automatically inserts `{}` and `;`.

### 2. put production unparenthesized into `pat`:

This allows one-of to appear:
- unparenthesized inside tuples: `f (3, one of {1;2}, 5) = 1`
- unparenthesized inside right sides of view-pats: `f (head -> one of {1;2}) = 3`

which feels more natural than having `()` around.

The rule is not too relaxed; this e.g. is still forbidden: `f one of 2;3 = 4`

### 3. or-pats with 0 or 1 arguments

This was requested by @eyeinski and seems like a nice thing to have. It may make sense to write an empty or-pattern and only later filling it with values.
Also, similar to `-XEmptyCase`, the value is evaluated before the pattern match fails.

As `f (one of) = 3` looks very weird, we require the syntax `one of {}` for empty or-patterns.
An or-pattern with a single alternative however could look like this: `f (one of 5) = 3`.

An empty or-pattern will give a redundant pattern match warning.