Amend #448 to avoid suspicious extension implication in GHC 9.8!
As written, #448 has a venerable old extension, `-XScopedTypeVariables,`
imply a brand-new extension `-XTypeAbstractions`. This was done for
backwards compat, but in light of the sentiment expressed in e.g. #601,
I think this is a bad mechanism for backwards compatibility.

There is a clear desire to separate more stable / less stable
extensions, and an implication like the above from stable to unstable
flagrantly violates it.

A new backwards compatibility mechanism is proposed instead.

I am sorry to say that I noticed this rather late, and the implication I
wish to change is ready to ship in **GHC 9.8**! As such, we need to
decide on this quickly.