Change semantics of `-fhpc` to not accumulate data over multiple runs (#612)
If I compile a program with the -fhpc option and run it, I expect to obtain a .tix file which contains information about the code covered during the execution of that run. This, however, is surprisingly not the behaviour that is implemented.
Instead, the generated .tix file contains the accumulated coverage of this execution of the program and all previous runs. I propose to change the semantics to only generate the coverage information for one run of the program.