Allow reserved identifiers as fields
This should enable the use of `HasField "type" s a` instances to be used with `OverloadedRecordDot` - `foo.type`.

This shows up for us in our database models. They include a type name prefix `data Foo = Foo { fooType :: FooType }`, and we have a TH step that removes that for `HasField` instances - `instance HasField "type" Foo FooType`. But we cannot write `foo.type` because it is a parse error.

Cf [this GHC ticket](https://gitlab.haskell.org/ghc/ghc/-/issues/24174), a change to the grammar here is necessary to make this change in general.